# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LearningSummary": {
        "additionalProperties": false,
        "properties": {
          "accuracy": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Accuracy",
            "type": "number"
          },
          "correct_questions": {
            "minimum": 0.0,
            "title": "Correct Questions",
            "type": "integer"
          },
          "course_id": {
            "minLength": 1,
            "title": "Course Id",
            "type": "string"
          },
          "due_review_count": {
            "minimum": 0.0,
            "title": "Due Review Count",
            "type": "integer"
          },
          "next_action": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Next Action",
            "type": "string"
          },
          "total_questions": {
            "minimum": 0.0,
            "title": "Total Questions",
            "type": "integer"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/LearningUnit"
            },
            "title": "Units",
            "type": "array"
          },
          "weak_units": {
            "items": {
              "$ref": "#/components/schemas/ReviewQueueItem"
            },
            "title": "Weak Units",
            "type": "array"
          }
        },
        "required": [
          "course_id",
          "total_questions",
          "correct_questions",
          "accuracy",
          "due_review_count",
          "next_action"
        ],
        "title": "LearningSummary",
        "type": "object"
      },
      "LearningUnit": {
        "additionalProperties": false,
        "properties": {
          "canonical_key": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Canonical Key",
            "type": "string"
          },
          "course_id": {
            "minLength": 1,
            "title": "Course Id",
            "type": "string"
          },
          "evidence_char_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Evidence Char Count",
            "type": "integer"
          },
          "evidence_chunk_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Evidence Chunk Count",
            "type": "integer"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/LearningUnitKind"
          },
          "label": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "mastery_level": {
            "$ref": "#/components/schemas/MasteryLevel",
            "default": "new"
          },
          "next_review_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review At"
          },
          "parent_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "practice_mode": {
            "$ref": "#/components/schemas/LearningUnitPracticeMode",
            "default": "knowledge_recall"
          },
          "practice_status": {
            "$ref": "#/components/schemas/LearningUnitPracticeStatus",
            "default": "insufficient_evidence"
          },
          "prototype_question_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuestionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/LearningUnitSource"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/LearningUnitStatus"
          }
        },
        "required": [
          "id",
          "course_id",
          "canonical_key",
          "label",
          "kind",
          "status"
        ],
        "title": "LearningUnit",
        "type": "object"
      },
      "LearningUnitKind": {
        "enum": [
          "section",
          "concept"
        ],
        "title": "LearningUnitKind",
        "type": "string"
      },
...
      "PracticeBatchRequest": {
        "additionalProperties": false,
        "properties": {
          "learning_unit_ids": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Learning Unit Ids",
            "type": "array"
          },
          "question_count": {
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Question Count",
            "type": "integer"
          }
        },
        "required": [
          "learning_unit_ids",
          "question_count"
        ],
        "title": "PracticeBatchRequest",
        "type": "object"
      },
...
      "PracticeSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "question_ids": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Question Ids",
            "type": "array"
          }
        },
        "required": [
          "question_ids"
        ],
        "title": "PracticeSessionRequest",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Serializable ``application/problem+json`` response body.",
        "properties": {
          "code": {
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code",
          "trace_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "期末复习智能体 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/learning-summary": {
      "get": {
        "operationId": "get_learning_summary_api_v1_courses__course_id__learning_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Learning Summary",
        "tags": [
          "learning-loop"
        ]
      }
    },
    "/api/v1/courses/{course_id}/learning-units": {
      "get": {
        "operationId": "list_learning_units_api_v1_courses__course_id__learning_units_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LearningUnit"
                  },
                  "title": "Response List Learning Units Api V1 Courses  Course Id  Learning Units Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Learning Units",
        "tags": [
          "learning-loop"
        ]
      }
    },
    "/api/v1/courses/{course_id}/learning-units/regenerate": {
      "post": {
        "operationId": "regenerate_learning_units_api_v1_courses__course_id__learning_units_regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LearningUnit"
                  },
                  "title": "Response Regenerate Learning Units Api V1 Courses  Course Id  Learning Units Regenerate Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Learning Units",
        "tags": [
          "learning-loop"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/practice-batches": {
      "post": {
        "operationId": "create_practice_batch_api_v1_courses__course_id__practice_batches_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeBatchSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求或来源状态冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "题目生成不可用"
          }
        },
        "summary": "Create Practice Batch",
        "tags": [
          "learning-loop"
        ]
      }
    },
    "/api/v1/courses/{course_id}/practice-sessions": {
      "post": {
        "operationId": "create_practice_session_api_v1_courses__course_id__practice_sessions_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSessionSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Practice Session",
        "tags": [
          "learning-loop"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/review-queue": {
      "get": {
        "operationId": "get_review_queue_api_v1_courses__course_id__review_queue_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewQueueItem"
                  },
                  "title": "Response Get Review Queue Api V1 Courses  Course Id  Review Queue Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Review Queue",
        "tags": [
          "learning-loop"
        ]
      }
    },
...
    "/api/v1/practice-batches/{batch_id}": {
      "get": {
        "operationId": "get_practice_batch_api_v1_practice_batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeBatchSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Practice Batch",
        "tags": [
          "learning-loop"
        ]
      }
    },
    "/api/v1/practice-sessions/{session_id}": {
      "get": {
        "operationId": "get_practice_session_api_v1_practice_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSessionSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Practice Session",
        "tags": [
          "learning-loop"
        ]
      }
    },
    "/api/v1/practice-sessions/{session_id}/attempts": {
      "post": {
        "operationId": "submit_practice_attempt_api_v1_practice_sessions__session_id__attempts_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeAttemptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeAttemptResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "题目或会话状态冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Practice Attempt",
        "tags": [
          "learning-loop"
        ]
      }
    },
    "/api/v1/practice-sessions/{session_id}/questions/{question_id}/tutor": {
      "get": {
        "operationId": "get_practice_tutor_conversation_api_v1_practice_sessions__session_id__questions__question_id__tutor_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeTutorConversation"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "题目来源状态冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Practice Tutor Conversation",
        "tags": [
          "learning-loop"
        ]
      },
      "post": {
        "operationId": "ask_practice_tutor_api_v1_practice_sessions__session_id__questions__question_id__tutor_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeTutorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeTutorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "题目来源状态冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "模型回答未通过约束"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "模型 Provider 不可用"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "模型响应超时"
          }
        },
        "summary": "Ask Practice Tutor",
        "tags": [
          "learning-loop"
        ]
      }
    },
...
  }
}